comments

+++ project2_scheduler/myscheduler/scheduler.c
@@ -35,6 +35,7 @@ struct queue {
     struct processNode *head, *tail;
 };
 
+// global pointer to the queue (it is global by nessecity)
 struct queue q = {
     NULL,
     NULL
@@ -48,18 +49,10 @@ struct processNode* newProcess(char *name){
     temp->next = NULL;
     temp->status = NEW;
     temp->prev = NULL;
+    // start timer
     clock_gettime(CLOCK_MONOTONIC, &temp->start_time);
     return temp;
 }
-
-// Initialize the queue
-struct queue* createQueue() {
-    struct queue* q 
-        = (struct queue*)malloc(sizeof(struct queue*));
-    q->head = NULL;
-    q->tail = NULL;
-    return q;
-};
 
 // Enqueue a process
 void enqueue(struct queue* q, char *name){
@@ -83,8 +76,9 @@ void enqueue(struct queue* q, char *name){
 
 // print info
 void print_info(struct processNode* k) {
-    //printf("Dequeue process with name %s\n", k->name);
+    // end timer
     clock_gettime(CLOCK_MONOTONIC, &k->end_time);
+    // calculate elapsed time
     int secs = k->end_time.tv_sec-k->start_time.tv_sec;
     int ms = (k->end_time.tv_nsec-k->start_time.tv_nsec)/1000;
     double res = (double)(secs + (double)ms/1000000);
@@ -104,7 +98,6 @@ void remove_from_queue(struct queue* q, pid_t pid){
     if (q->head == q->tail) {
         print_info(q->head);
         q->head->next = q->head->prev = NULL;
-        // TODO////////////////////////////////CHANGE TO WORKLOAD TIME
         printf("WORKLOAD TIME: %.2f secs\n", q->head->elapsed_time);
         free(q->head);
         q->head = q->tail = NULL;
@@ -113,6 +106,7 @@ void remove_from_queue(struct queue* q, pid_t pid){
 
     struct processNode* temp = q->head;
 
+    // find the node with given pid
     while (temp->pid != pid && temp != q->tail) {
         temp = temp->next;
     }
@@ -138,7 +132,9 @@ void remove_from_queue(struct queue* q, pid_t pid){
     return;
 }
 
-void execute(struct processNode* process) {
+// execute the process
+void execute(struct processNode* process, struct timespec request,
+             struct timespec remaining) {
 
     pid_t pid; 
 
@@ -149,14 +145,18 @@ void execute(struct processNode* process) {
             perror("fork");
             exit(EXIT_FAILURE);
         }
+        // child
         else if (pid == 0) {
             execlp(process->name, process->name, NULL);
             exit(EXIT_SUCCESS);
         }
+        // parent
         else {
             printf("executing %s\n", process->name);
             process->pid = pid;
             process->status = RUNNING;
+            // wait the quantum
+            nanosleep(&request, &remaining);
         }
     }
     // process was stopped
@@ -164,22 +164,30 @@ void execute(struct processNode* process) {
         printf("executing %s\n", process->name);
         kill(process->pid, SIGCONT);
         process->status = RUNNING;
+        // wait the quantum
+        nanosleep(&request, &remaining);
     }
 }
 
+// rr scheduler, quantum is given in miliseconds 
+// (fcfs is performed with very big quantum)
 void rr_scheduler(struct queue* q, unsigned long long quantum) {
 
     struct processNode* k = q->head;
     struct timespec remaining, request;
+    // calculate tv_sec and tv_nsec
     request.tv_sec = quantum/1000;
     quantum = quantum%1000;
     request.tv_nsec = quantum * 1000000;
 
+    // schedule every process in order until the list is empty
     while (k != NULL) {
-        execute(k);
-        nanosleep(&request, &remaining);
-        kill(k->pid, SIGSTOP);
-        k->status = STOPPED;
+        execute(k, request, remaining);
+        // only stop the process if its not blocked
+        if (k->status == RUNNING) {
+            kill(k->pid, SIGSTOP);
+            k->status = STOPPED;
+        }
         k = k->next;
     }; 
 
@@ -187,19 +195,22 @@ void rr_scheduler(struct queue* q, unsigned long long quantum) {
 
 }
 
+// handle SIGCHLD
 void sigchld_handler(int signo, siginfo_t *si, void *unused) {
     remove_from_queue(&q, si->si_pid);
 }
 
+// putting it all together
 int main(int argc, char **argv) {
 
     FILE *fp;
     char *line = NULL;
     size_t len = 0;
     ssize_t readline;
-    int i, lines = 0;
+    int lines = 0;
     char ch;
 
+    // set sigaction for SIGCHLD
     struct sigaction sachld;                                        
     sigemptyset( &sachld.sa_mask );                                 
     sachld.sa_flags = SA_SIGINFO | SA_RESTART | SA_NOCLDSTOP;
@@ -219,23 +230,17 @@ int main(int argc, char **argv) {
         exit(EXIT_FAILURE);
 
     // count number of lines in FILE
-    while(!feof(fp))
-    {
+    while(!feof(fp)) {
         ch = fgetc(fp);
         if(ch == '\n')
-        {
             lines++;
-        }
     }
 
     // reset pointer to start of FILE
     rewind(fp);
 
-    // Initiate queue
-
-    i = -1;
+    // populate queue
     while ((readline = getline(&line, &len, fp)) != -1) {
-        i++;
         //remove trailing newline
         if (line[strlen(line) - 1] == '\n')
             line[strlen(line) - 1] = '\0';
@@ -245,14 +250,13 @@ int main(int argc, char **argv) {
     q.tail->next = q.head;
     q.head->prev = q.tail;
 
-    for (int i=0; i<lines; i++){
-        wait(NULL);
-    }
     //free(line);
     fclose(fp);
 
+    // fcfs (quantum is very big)
     if (argc == 3)
         rr_scheduler(&q, ULLONG_MAX);
+    // round robin with given quantum
     else
         rr_scheduler(&q, strtoull(argv[2], NULL, 10));
